+++ src/main/java/com/example/carrentalproject/repository/OrderRepository.java
@@ -1,4 +1,4 @@
-package com.example.carrentalproject.repo;
+package com.example.carrentalproject.repository;
 
 import com.example.carrentalproject.domain.PlacedOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
