+++ src/main/java/com/example/carrentalproject/service/CarService.java
@@ -5,9 +5,9 @@ import com.example.carrentalproject.domain.CarPackage;
 import com.example.carrentalproject.domain.CarParameters;
 import com.example.carrentalproject.dto.CarDto;
 import com.example.carrentalproject.dto.CarPackageDto;
-import com.example.carrentalproject.repo.CarPackageRepository;
-import com.example.carrentalproject.repo.CarParametersRepository;
-import com.example.carrentalproject.repo.CarRepository;
+import com.example.carrentalproject.repository.CarPackageRepository;
+import com.example.carrentalproject.repository.CarParametersRepository;
+import com.example.carrentalproject.repository.CarRepository;
 import com.example.carrentalproject.exception.ExistingEntityException;
 import com.example.carrentalproject.utils.PageValidator;
 import com.example.carrentalproject.utils.SortValidator;
