+++ src/main/java/com/example/carrentalproject/service/DeliveryService.java
@@ -4,12 +4,12 @@ import com.example.carrentalproject.domain.Car;
 import com.example.carrentalproject.domain.PlacedOrder;
 import com.example.carrentalproject.domain.User;
 import com.example.carrentalproject.exception.UnavailableCarException;
-import com.example.carrentalproject.repo.CarRepository;
-import com.example.carrentalproject.repo.OrderRepository;
+import com.example.carrentalproject.repository.CarRepository;
+import com.example.carrentalproject.repository.OrderRepository;
 import com.example.carrentalproject.security.LoggedInUser;
 import com.example.carrentalproject.exception.InvalidPackageException;
 import com.example.carrentalproject.exception.NoAccessKeyException;
-import com.example.carrentalproject.repo.AccessKeyRepository;
+import com.example.carrentalproject.repository.AccessKeyRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
