+++ src/main/java/com/example/carrentalproject/service/OrderService.java
@@ -6,12 +6,12 @@ import com.example.carrentalproject.domain.PlacedOrder;
 import com.example.carrentalproject.domain.User;
 import com.example.carrentalproject.dto.AccessKeyDto;
 import com.example.carrentalproject.exception.ExistingOrderException;
-import com.example.carrentalproject.repo.CarPackageRepository;
-import com.example.carrentalproject.repo.OrderRepository;
+import com.example.carrentalproject.repository.CarPackageRepository;
+import com.example.carrentalproject.repository.OrderRepository;
 import com.example.carrentalproject.security.LoggedInUser;
 import com.example.carrentalproject.exception.InsufficientFundsException;
 import com.example.carrentalproject.exception.NoCreditCardException;
-import com.example.carrentalproject.repo.AccessKeyRepository;
+import com.example.carrentalproject.repository.AccessKeyRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
