+++ src/main/java/com/example/carrentalproject/service/PaymentService.java
@@ -4,8 +4,8 @@ import com.example.carrentalproject.domain.CreditCard;
 import com.example.carrentalproject.domain.User;
 import com.example.carrentalproject.dto.CreditCardDto;
 import com.example.carrentalproject.exception.NoCreditCardException;
-import com.example.carrentalproject.repo.CreditCardRepository;
-import com.example.carrentalproject.repo.UserRepository;
+import com.example.carrentalproject.repository.CreditCardRepository;
+import com.example.carrentalproject.repository.UserRepository;
 import com.example.carrentalproject.security.LoggedInUser;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
