+++ src/main/java/com/example/carrentalproject/service/UserService.java
@@ -7,13 +7,13 @@ import com.example.carrentalproject.dto.CreditCardDto;
 import com.example.carrentalproject.dto.UserDto;
 import com.example.carrentalproject.mapper.UserDtoMapper;
 import com.example.carrentalproject.mapper.UserInDtoMapper;
-import com.example.carrentalproject.repo.CreditCardRepository;
-import com.example.carrentalproject.repo.RoleRepository;
+import com.example.carrentalproject.repository.CreditCardRepository;
+import com.example.carrentalproject.repository.RoleRepository;
 import com.example.carrentalproject.dto.UserInDto;
 import com.example.carrentalproject.exception.AssignedRoleException;
 import com.example.carrentalproject.exception.ExistingEntityException;
 import com.example.carrentalproject.exception.NoCreditCardException;
-import com.example.carrentalproject.repo.UserRepository;
+import com.example.carrentalproject.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
