+++ src/test/java/com/example/carrentalproject/repository/RoleRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.example.carrentalproject.repo;
+package com.example.carrentalproject.repository;
 
 import com.example.carrentalproject.domain.Role;
 import org.assertj.core.api.AssertionsForClassTypes;
