+++ src/test/java/com/example/carrentalproject/repository/UserRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.example.carrentalproject.repo;
+package com.example.carrentalproject.repository;
 
 import com.example.carrentalproject.domain.User;
 import org.assertj.core.api.AssertionsForClassTypes;
