+++ src/test/java/com/example/carrentalproject/service/OrderServiceTest.java
@@ -8,9 +8,9 @@ import com.example.carrentalproject.domain.User;
 import com.example.carrentalproject.dto.AccessKeyDto;
 import com.example.carrentalproject.exception.InsufficientFundsException;
 import com.example.carrentalproject.exception.NoCreditCardException;
-import com.example.carrentalproject.repo.AccessKeyRepository;
-import com.example.carrentalproject.repo.CarPackageRepository;
-import com.example.carrentalproject.repo.OrderRepository;
+import com.example.carrentalproject.repository.AccessKeyRepository;
+import com.example.carrentalproject.repository.CarPackageRepository;
+import com.example.carrentalproject.repository.OrderRepository;
 import com.example.carrentalproject.security.LoggedInUser;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
