+++ src/test/java/com/example/carrentalproject/service/UserServiceTest.java
@@ -8,9 +8,9 @@ import com.example.carrentalproject.dto.UserInDto;
 import com.example.carrentalproject.exception.AssignedRoleException;
 import com.example.carrentalproject.exception.ExistingEntityException;
 import com.example.carrentalproject.mapper.UserDtoMapper;
-import com.example.carrentalproject.repo.CreditCardRepository;
-import com.example.carrentalproject.repo.RoleRepository;
-import com.example.carrentalproject.repo.UserRepository;
+import com.example.carrentalproject.repository.CreditCardRepository;
+import com.example.carrentalproject.repository.RoleRepository;
+import com.example.carrentalproject.repository.UserRepository;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
